Split edit secret into separate components for each type
