+++ edit/secret/cloud.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>Cloud</div>
+</template>
